Fix location search in maps view
git-svn-id: http://redbus.googlecode.com/svn/trunk@354 047a1783-e752-e205-4d80-9245024c4953

+++ src/org/redbus/GeocodingHelper.java
@@ -50,7 +50,7 @@ public class GeocodingHelper {
 			
 			try {
 				Geocoder geocoder = new Geocoder(gr.ctx, Locale.UK);
-				gr.addresses = geocoder.getFromLocationName(gr.location, 5, pt.lowerLeftLat / 1E6, pt.lowerLeftLon / 1E6, pt.upperRightLat / 1E6, pt.upperRightLon / 1E6);
+				gr.addresses = geocoder.getFromLocationName(gr.location, 5, (double) pt.lowerLeftLat / 1E6, (double) pt.lowerLeftLon / 1E6, (double) pt.upperRightLat / 1E6, (double) pt.upperRightLon / 1E6);
 			} catch (Throwable t) {
 				Log.e("AsyncHttpRequestTask.doInBackGround", "Throwable", t);
 			}
